refactor to include an actual test
